+++ Network/Telegram/API/Bot/Object/Member.hs
@@ -4,6 +4,7 @@ import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
 import "aeson" Data.Aeson.Types (Object, Parser)
 import "base" Control.Applicative (Alternative ((<|>)))
 
+import Network.Telegram.API.Bot.Has (Has (focus))
 import Network.Telegram.API.Bot.Object.Chat (Chat)
 import Network.Telegram.API.Bot.Object.From (From)
 
@@ -17,3 +18,7 @@ instance FromJSON Member where
 
 		joined :: Object -> Parser Member
 		joined v = Joined <$> v .: "chat" <*> v .: "new_chat_members"
+
+instance Has Member Chat where
+	focus f (Gone chat from) = (\chat' -> Gone chat' from) <$> f chat
+	focus f (Joined chat froms) = (\chat' -> Joined chat' froms) <$> f chat