+++ Network/Telegram/API/Bot/Object/Update.hs
@@ -3,7 +3,6 @@ module Network.Telegram.API.Bot.Object.Update (Update (..)) where
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
 import "aeson" Data.Aeson.Types (Object, Parser)
 import "base" Control.Applicative (Alternative ((<|>)))
-import "lens" Control.Lens (Lens')
 
 import Network.Telegram.API.Bot.Has (Has (focus))
 import Network.Telegram.API.Bot.Object.Callback (Callback (Datatext))
@@ -40,15 +39,9 @@ instance FromJSON Update where
 		incoming v = Incoming <$> v .: "update_id" <*> v .: "message"
 
 instance Has Update Chat where
-	focus f (Incoming upd_id (Command msg_id chat_ from cmd)) =
-		(\chat' -> Incoming upd_id (Command msg_id chat' from cmd)) <$> f chat_
-	focus f (Incoming upd_id (Textual msg_id chat_ from txt)) =
-		(\chat' -> Incoming upd_id (Textual msg_id chat' from txt)) <$> f chat_
-	focus f (Query upd_id (Datatext cq_id (Command msg_id chat_ from cmd) dttxt)) =
-		(\chat' -> Query upd_id (Datatext cq_id (Command msg_id chat' from cmd) dttxt)) <$> f chat_
-	focus f (Query upd_id (Datatext cq_id (Textual msg_id chat_ from txt) dttxt)) =
-		(\chat' -> Query upd_id (Datatext cq_id (Textual msg_id chat' from txt) dttxt)) <$> f chat_
-	focus f (Membership upd_id (Gone chat_ users)) =
-		(\chat' -> Membership upd_id (Gone chat' users)) <$> f chat_
-	focus f (Membership upd_id (Joined chat_ users)) =
-		(\chat' -> Membership upd_id (Joined chat' users)) <$> f chat_
+	focus f (Incoming upd_id (Command msg_id chat_ from cmd)) = (\chat' -> Incoming upd_id (Command msg_id chat' from cmd)) <$> f chat_
+	focus f (Incoming upd_id (Textual msg_id chat_ from txt)) = (\chat' -> Incoming upd_id (Textual msg_id chat' from txt)) <$> f chat_
+	focus f (Query upd_id (Datatext cq_id (Command msg_id chat_ from cmd) dttxt)) = (\chat' -> Query upd_id (Datatext cq_id (Command msg_id chat' from cmd) dttxt)) <$> f chat_
+	focus f (Query upd_id (Datatext cq_id (Textual msg_id chat_ from txt) dttxt)) = (\chat' -> Query upd_id (Datatext cq_id (Textual msg_id chat' from txt) dttxt)) <$> f chat_
+	focus f (Membership upd_id (Gone chat_ users)) = (\chat' -> Membership upd_id (Gone chat' users)) <$> f chat_
+	focus f (Membership upd_id (Joined chat_ users)) = (\chat' -> Membership upd_id (Joined chat' users)) <$> f chat_